frontend/requirements.txt
#frontend/requirements.txt
streamlit
requests
python-dotenv